CategoryController ++

+++ ewm-service/src/main/java/ru/practicum/ewm_main_service/category/CategoryController.java
@@ -7,8 +7,9 @@ import org.springframework.web.bind.annotation.*;
 import ru.practicum.ewm_main_service.category.dto.CategoryDto;
 import ru.practicum.ewm_main_service.category.dto.NewCategoryDto;
 import ru.practicum.ewm_main_service.category.service.CategoryService;
+import ru.practicum.ewm_main_service.event.service.EventService;
+import ru.practicum.ewm_main_service.exception.DataAlreadyExists;
 import ru.practicum.ewm_main_service.exception.ValidationException;
-import ru.practicum.ewm_main_service.user.dto.UserDto;
 
 import java.util.List;
 
@@ -17,6 +18,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CategoryController {
     private final CategoryService categoryService;
+    private final EventService eventService;
 
     @PostMapping("/admin/categories")
     @ResponseStatus(HttpStatus.CREATED)
@@ -54,7 +56,9 @@ public class CategoryController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteCategoryById(@PathVariable Long catId) {
         log.info("Получен запрос - удалить данные категории '{}'", catId);
+        if (!eventService.findEventByCategoryId(catId).isEmpty()) {
+            throw new DataAlreadyExists(String.format("Существуют события, связанные с категорией %s", catId));
+        }
         categoryService.deleteCategoryById(catId);
-        //Обратите внимание: с категорией не должно быть связано ни одного события.
     }
 }
